requirements.txt
langchain==0.3.1
langchain-community
langchain-groq
langchain-core
langchain-cli
python-dotenv
sentence_transformers
docarray
streamlit==1.24.0 
streamlit-chat
uvicorn
fastapi
pandas
altair<5
